+++ package/src/components/AudioPlayer/Interface/playListPortalContext.ts
@@ -1,4 +1,13 @@
 import { createContext } from "react";
+import { useNonNullableContext } from "@/hooks/context/useNonNullableContext";
 
-export const playListPortalContext = createContext<HTMLDivElement | null>(null);
+export interface PlayListPortalContextValue {
+  node: HTMLDivElement | null;
+}
+
+export const playListPortalContext =
+  createContext<PlayListPortalContextValue | null>(null);
 playListPortalContext.displayName = "PlayListPortalContext";
+
+export const usePlayListPortalContext = () =>
+  useNonNullableContext(playListPortalContext);